requirements.txt
torch==2.0.1
torchvision==0.15.2
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
Pillow>=9.5.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
